Successfully tested Oye() with a instead of vind
Stole a lot of Carlos' code, but he uses vind instead of a, so I to
rewrite for a. There was a lot of struggle with his obsession with the
negative, but it appears to be working now. There is still the issue
of needing to input the quasi-steady CT when we implement it properly,
but I have written about this in utilities_DI.